fix "validateRequest" (transfer of logic to the service)
