+++ src/service/users.service.ts
@@ -1,6 +1,8 @@
 import UsersDao from "../daos/users.dao";
 import { IUser } from "../model/userModel";
 import bcrypt from "bcryptjs";
+import Joi from "@hapi/joi";
+import { NextFunction } from "express";
 
 class UsersService {
   async create(resource: IUser) {
@@ -13,6 +15,26 @@ class UsersService {
 
   async authenticate(login: string, password: string) {
     return UsersDao.authenticate_function(login, password);
+  }
+
+  async validateRequest(login: string, password: string, next: NextFunction) {
+    const options = {
+      abortEarly: false,
+      allowUnknown: true,
+      stripUnknown: true,
+    };
+
+    const schema = Joi.object({
+      login: Joi.string().required(),
+      password: Joi.string().required(),
+    });
+
+    const { error } = schema.validate({ login, password }, options);
+
+    if (error) {
+      throw error;
+    }
+    next();
   }
 }
